+++ src/redux/sagas/bookSaga.js
@@ -1,0 +1,56 @@
+import { takeEvery, put } from 'redux-saga/effects';
+import axios from 'axios';
+
+
+
+function* fetchBooks(action) {
+    try {
+        const serverResponse = yield axios.get('/api/addbooks');
+        const action = {
+            type: 'SET_BOOKS',
+            payload: serverResponse.data
+        };
+        yield put(action); // triggers the reducer
+    } catch (error) {
+        console.log('Error in axios GET: ', error);
+        alert('Something Went Wrong!');
+    }
+}
+
+function* postBooks(action) {
+    try {
+        yield axios.post('/api/addbooks', action.payload);
+        console.log('action.payload: ', action.payload);
+
+        const nextAction = {
+            type: 'FETCH_BOOKS'
+        };
+        yield put(nextAction);
+    } catch (error) {
+        console.log('Error in POST');
+        alert('There is a problem in POST');
+    }
+}
+
+function* deleteBooks(action) {
+    try {
+        yield axios.delete(`/api/addbooks/${action.payload}`);
+        console.log('action.payload ', action.payload);
+
+        const nextAction = {
+            type: 'FETCH_BOOKS'
+        };
+        yield put(nextAction);
+    } catch (error) {
+        console.log('Error in DELETE: ', error);
+        alert('Something went wrong')
+    }
+}
+
+function* bookSaga() {
+    yield takeEvery('FETCH_BOOKS', fetchBooks);
+    yield takeEvery('ADD_BOOKS', postBooks);
+    yield takeEvery('DELETE_BOOKS', deleteBooks)
+}
+
+export default bookSaga;
